rebuild/pyproject.toml
[project]
name = "rebuild"
version = "0.0.0"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi >=0.115,<0.116",
  "uvicorn[standard] >=0.32,<0.33",
  "sqlalchemy[asyncio] >=2.0.36,<2.1",
  "asyncmy >=0.2.10,<0.3",
  "cryptography >=43,<46",          # required by asyncmy for MySQL 8 caching_sha2_password
  "alembic >=1.13,<1.14",
  "pydantic >=2.9,<3",
  "pydantic-settings >=2.6,<3",
  "python-socketio >=5.11,<6",     # installed for M3, unused in M0
  "apscheduler >=3.10,<4",          # installed for M4, unused in M0
  "openai >=1.55,<2",               # installed for M1, unused in M0
  "httpx >=0.27,<0.29",
  "redis >=5.1,<6",
  "uuid7-standard >=1.1,<2",        # RFC 9562 UUIDv7 backport (Python 3.12); see app/core/ids.py
  "boto3 >=1.35,<2",                # AWS RDS IAM database authentication; imported lazily inside app.core.iam_auth
]

[dependency-groups]
dev = [
  "ruff >=0.7.4,<0.8",
  "mypy >=1.13,<2",
  "pytest >=8.3,<9",
  "pytest-asyncio >=0.24,<0.25",
  "pytest-cov >=5,<6",
  "anyio >=4.6,<5",
  "testcontainers[mysql] >=4.8,<5",
  "fakeredis >=2.26,<3",
  "boto3-stubs[rds] >=1.35,<2",     # IAM auth helper is typed against RDSClient; mypy strict needs the stubs
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/app"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["backend"]
extend-exclude = [".venv", "node_modules", "frontend"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "ASYNC", "RET", "PL", "PT", "TID"]
ignore = ["PLR0913"]

[tool.ruff.lint.per-file-ignores]
"alembic/versions/*.py" = ["E501"]
# Alembic env.py defers `app.*` imports until call time on purpose so
# the module remains importable in isolation (docstring at the top of
# the file). PLC0415 ("import outside top-level") would otherwise fire
# on every `_settings_url` / `_target_metadata` / `run_async_migrations`.
"alembic/env.py" = ["PLC0415"]
"tests/*.py" = ["PLR2004", "TID251"]

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"uuid.uuid4".msg = "Use app.core.ids.new_id() (UUIDv7) instead. See rebuild.md §9."
"datetime.datetime.utcnow".msg = "Use app.core.time.now_ms() (epoch-ms BIGINT) instead. See rebuild.md §4."

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_unreachable = true
disallow_any_explicit = false  # SQLAlchemy generics force a few Any escapes
plugins = ["pydantic.mypy"]
files = ["backend/app"]

[[tool.mypy.overrides]]
module = ["asyncmy.*", "alembic.*", "uuid7"]
ignore_missing_imports = true

# boto3 ships partial stubs (botocore types are strict, the high-level
# `boto3.client(...)` returns a generic client). The IAM auth helper
# only calls `client.generate_db_auth_token(...)` which is well-typed
# via `boto3-stubs[rds]`, but we ignore_missing_imports for the lazy
# import path inside `app.core.iam_auth.generate_iam_auth_token` so a
# dev environment without the stubs installed still typechecks.
[[tool.mypy.overrides]]
module = ["boto3", "boto3.*", "botocore", "botocore.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-ra -q --strict-markers --strict-config"
asyncio_mode = "auto"
# pytest-asyncio 0.24: pin fixture loop scope explicitly to silence the
# "unset default loop scope" deprecation warning. We keep it at "function"
# (the historical default) so async fixtures share a loop with their test —
# the test ``engine`` fixture uses ``NullPool`` so connections are never
# retained across loops, which removes the cross-loop binding hazard that
# would otherwise force a session-scoped loop.
asyncio_default_fixture_loop_scope = "function"
testpaths = ["backend/tests"]
filterwarnings = ["error::DeprecationWarning"]
